pyproject.toml
[tool.poetry]
name = "onecode"
version = "0.1.0"
description = "Python skeleton and library for OneCode projects"
readme = "README.md"
authors = ["DeepLime <contact@deeplime.io>"]
license = "MIT"
keywords = ["onecode", "procedures"]
homepage = "https://github.com/deeplime-io/onecode"
repository = "https://github.com/deeplime-io/onecode"
documentation = "https://github.com/deeplime-io/onecode"
classifiers = [
  "Development Status :: 3 - Beta",
  "Intended Audience :: Developers",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Natural Language :: English",
  "Operating System :: Unix",
  "Operating System :: Microsoft :: Windows",
  "Programming Language :: Python",
  "Topic :: Scientific/Engineering",
  "Topic :: Software Development :: Code Generators"
]

[tool.poetry.dependencies]
python = ">=3.7, <3.11"
astunparse = "^1.6.3"
'flufl.lock' = "^7.1.1"
inquirerpy = "^0.3.3"
pandas = "^1.4.0"
pyarrow = "^6.0.0"
pydantic = "^1.10.4"
pycg = { git = "https://github.com/deeplime-io/PyCG.git@onecode" }
pydash = "^5.1.0"
python-slugify = "^6.1.2"
typeguard = "^2.13.3"
yaspin = "^2.1.0"

# streamlit
streamlit = { version = ">=1.12.0,<1.18", optional = true }
streamlit-image-select = { version = "^0.5.1", optional = true }
streamlit-option-menu = { version = "^0.3.2", optional = true }
streamlit_tree_select = { version = "^0.0.5", optional = true }

# docs
griffe = { version = "==0.25.0", optional = true }
markdown-katex = { version = "==202112.1034", optional = true }
mike = { version = "==1.1.2", optional = true }
mkdocs = { version = "==1.4.2", optional = true }
mkdocs-jupyter = { version = "==0.21.0", optional = true }
mkdocs-material = { version = "==8.5.10", optional = true }
mkdocstrings = { version = "==0.19.0", optional = true }
mkdocstrings-python = { version = "==0.7.1", optional = true }

# tests
datatest = { version = "^0.11.1", optional = true }
pytest = { version = "^6.0.2", optional = true }
pytest-cov = { version = "^2.10.1", optional = true }
pytest-xdist = { version = "^3.1.0", optional = true }
toml = { version = "^0.10.2", optional = true }

[tool.poetry.extras]
tech-expert = ["streamlit", "streamlit-image-select", "streamlit-option-menu", "streamlit_tree_select"]
developer = ["streamlit", "streamlit-image-select", "streamlit-option-menu", "streamlit_tree_select", "datatest", "pytest", "pytest-cov", "pytest-xdist", "toml"]
docs = ["griffe", "markdown-katex", "mike", "mkdocs", "mkdocs-jupyter", "mkdocs-material", "mkdocstrings", "mkdocstrings-python"]

[tool.poetry.scripts]
onecode-start = 'onecode.cli.start:main'
onecode-create = 'onecode.cli.create:main'
onecode-add = 'onecode.cli.add:main'
onecode-extract = 'onecode.cli.extract:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
